Local agent sessions provider cleanup (#279359)
